+++ server/api/Boltz.ts
@@ -6,8 +6,8 @@ export const BoltzApi = {
   getPairs: async () => {
     try {
       const response = await fetchWithProxy(`${appUrls.boltz}/getpairs`);
-      return await response.json();
-    } catch (error) {
+      return (await response.json()) as any;
+    } catch (error: any) {
       logger.error('Error getting pairs from Boltz: %o', error);
       throw new Error('ErrorGettingBoltzPairs');
     }
@@ -17,8 +17,8 @@ export const BoltzApi = {
       const response = await fetchWithProxy(
         `${appUrls.boltz}/getfeeestimation`
       );
-      return await response.json();
-    } catch (error) {
+      return (await response.json()) as any;
+    } catch (error: any) {
       logger.error('Error getting fee estimations from Boltz: %o', error);
       throw new Error(error);
     }
@@ -31,8 +31,8 @@ export const BoltzApi = {
         body: JSON.stringify(body),
         headers: { 'Content-Type': 'application/json' },
       });
-      return await response.json();
-    } catch (error) {
+      return (await response.json()) as any;
+    } catch (error: any) {
       logger.error('Error getting fee estimations from Boltz: %o', error);
       throw new Error(error);
     }
@@ -56,8 +56,8 @@ export const BoltzApi = {
         body: JSON.stringify(body),
         headers: { 'Content-Type': 'application/json' },
       });
-      return await response.json();
-    } catch (error) {
+      return (await response.json()) as any;
+    } catch (error: any) {
       logger.error('Error getting fee estimations from Boltz: %o', error);
       throw new Error(error);
     }
@@ -76,8 +76,8 @@ export const BoltzApi = {
           headers: { 'Content-Type': 'application/json' },
         }
       );
-      return await response.json();
-    } catch (error) {
+      return (await response.json()) as any;
+    } catch (error: any) {
       logger.error('Error broadcasting transaction from Boltz: %o', error);
       throw new Error(error);
     }