+++ server/schema/lnurl/resolvers.ts
@@ -84,7 +84,7 @@ export const lnUrlResolvers = {
 
       try {
         const response = await fetchWithProxy(finalUrl);
-        const json = await response.json();
+        const json = (await response.json()) as any;
 
         logger.debug('LnUrlAuth response: %o', json);
 
@@ -93,7 +93,7 @@ export const lnUrlResolvers = {
         }
 
         return { ...json, message: json.event || 'LnServiceSuccess' };
-      } catch (error) {
+      } catch (error: any) {
         logger.error('Error authenticating with LnUrl service: %o', error);
         throw new Error('ProblemAuthenticatingWithLnUrlService');
       }
@@ -107,14 +107,14 @@ export const lnUrlResolvers = {
 
       try {
         const response = await fetchWithProxy(url);
-        const json = await response.json();
+        const json = (await response.json()) as any;
 
         if (json.status === 'ERROR') {
           throw new Error(json.reason || 'LnServiceError');
         }
 
         return json;
-      } catch (error) {
+      } catch (error: any) {
         logger.error('Error fetching from LnUrl service: %o', error);
         throw new Error('ProblemFetchingFromLnUrlService');
       }
@@ -149,12 +149,12 @@ export const lnUrlResolvers = {
 
       try {
         const response = await fetchWithProxy(finalUrl);
-        lnServiceResponse = await response.json();
+        lnServiceResponse = (await response.json()) as any;
 
         if (lnServiceResponse.status === 'ERROR') {
           throw new Error(lnServiceResponse.reason || 'LnServiceError');
         }
-      } catch (error) {
+      } catch (error: any) {
         logger.error('Error paying to LnUrl service: %o', error);
         throw new Error('ProblemPayingLnUrlService');
       }
@@ -231,7 +231,7 @@ export const lnUrlResolvers = {
 
       try {
         const response = await fetchWithProxy(finalUrl);
-        const json = await response.json();
+        const json = (await response.json()) as any;
 
         logger.debug('LnUrlWithdraw response: %o', json);
 
@@ -241,7 +241,7 @@ export const lnUrlResolvers = {
 
         // Return invoice id to check status
         return info.id;
-      } catch (error) {
+      } catch (error: any) {
         logger.error('Error withdrawing from LnUrl service: %o', error);
         throw new Error('ProblemWithdrawingFromLnUrlService');
       }
@@ -273,7 +273,7 @@ export const lnUrlResolvers = {
 
       try {
         const response = await fetchWithProxy(finalUrl);
-        const json = await response.json();
+        const json = (await response.json()) as any;
 
         logger.debug('LnUrlChannel response: %o', json);
 
@@ -282,7 +282,7 @@ export const lnUrlResolvers = {
         }
 
         return 'Successfully requested a channel open';
-      } catch (error) {
+      } catch (error: any) {
         logger.error('Error requesting channel from LnUrl service: %o', error);
         throw new Error(
           `Error requesting channel from LnUrl service: ${error}`